+++ services/els-verify/src/main/java/app/coronawarn/datadonation/services/els/otp/ElsOtpControllerExceptionHandler.java
@@ -1,4 +1,4 @@
-package app.coronawarn.datadonation.services.edus.otp;
+package app.coronawarn.datadonation.services.els.otp;
 
 import app.coronawarn.datadonation.common.persistence.service.OtpNotFoundException;
 import org.slf4j.Logger;
