+++ services/els-verify/src/test/java/app/coronawarn/datadonation/services/els/otp/GenerateElsOtpControllerTest.java
@@ -1,4 +1,4 @@
-package app.coronawarn.datadonation.services.edus.otp;
+package app.coronawarn.datadonation.services.els.otp;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -13,7 +13,7 @@ import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.ActiveProfiles;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@ActiveProfiles("generate-els-otp")
+@ActiveProfiles("generate-els")
 @DirtiesContext
 public class GenerateElsOtpControllerTest {
